+++ Bob'n'Eve/EntityComposite.h
@@ -6,10 +6,10 @@
 class EntityComposite : public Entity
 {
 public:
-	EntityComposite(float layer, int id, Entity::TypeEnum type);
+	EntityComposite(float layer, int32_t id, Entity::TypeEnum type);
 	virtual ~EntityComposite();
 
-	void Update(int ms) override;
+	void Update(int16_t ms) override;
 	void Draw(Renderer& renderer) override;
 
 	void Add(Entity* e);
